VirtualConnection improvements round2
